Thêm dto, repo, services và fix lỗi

+++ backend/src/main/java/com/khai/admin/service/ProductService.java
@@ -1,8 +1,13 @@
 package com.khai.admin.service;
 
+import com.khai.admin.dto.ProductDto;
+import com.khai.admin.dto.user.UserView;
 import com.khai.admin.entity.Category;
 import com.khai.admin.entity.Product;
+import com.khai.admin.entity.User;
 import com.khai.admin.exception.AlreadyExist;
+import com.khai.admin.exception.NoSuchElementException;
+import com.khai.admin.repository.CategoryRepository;
 import com.khai.admin.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
@@ -18,16 +23,24 @@ import java.util.*;
 @Service
 public class ProductService {
     private ProductRepository productRepository;
+    private UserService userService;
+    private final CategoryRepository categoryRepository;
 
     @Autowired
-    public ProductService(ProductRepository productRepository) {
+    public ProductService(
+        ProductRepository productRepository,
+        CategoryRepository categoryRepository,
+        UserService userService
+    ) {
         this.productRepository = productRepository;
+        this.categoryRepository = categoryRepository;
+        this.userService = userService;
     }
 
-    public Product getProductInfo(int id) {
+    public ProductDto getProductInfo(int id) {
         Optional<Product> product = productRepository.findById(id);
         if(product.isPresent()) {
-            return product.get();
+            return product.get().toDto();
         }
         throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Không tìm thấy sản phẩm có id" + id);
     }
@@ -51,21 +64,33 @@ public class ProductService {
         }
     }
 
-    public Product create(Product product) {
-        Date now = new Date();
-        if(productRepository.isExist(product.getName(), product.getCode()).isPresent()) {
-            throw new AlreadyExist("Sản phẩm", product.getName());
-        }
-        product.setDeleted(false);
-        product.setCreatedDate(now);
-//        product.setLast_modified(now);
+    @Transactional
+    public ProductDto create(Map<String, String> headers, Product product) {
+        try {
+            Date now = new Date();
+            if(productRepository.isExist(product.getName(), product.getCode()).isPresent()) {
+                throw new AlreadyExist("Sản phẩm");
+            }
+            User user = userService.getUserById(1);
 
-        return productRepository.save(product);
+            product.setDeleted(false);
+            product.setCreatedDate(now);
+            Category category = categoryRepository.findById(product.getCategory().getId()).orElse(null);
+            if(category != null)
+                product.setCategory(category);
+            product.setCreator(user);
+            productRepository.save(product);
+
+            return product.toDto();
+        } catch(Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
+        }
     }
 
     public Map<String, Object> getProducts(String name, Pageable pageable) {
         try {
             List<Product> products = new ArrayList<>();
+            List<ProductDto> productDtoList = new ArrayList<>();
             Page<Product> pageProducts;
 //            if(name == null) {
                 pageProducts = productRepository.findAll(pageable);
@@ -74,8 +99,9 @@ public class ProductService {
 //                pageProducts = productRepository.findByNameContaining(name, pageable);
 //            }
             products = pageProducts.getContent();
+            products.stream().map(ProductDto::new).forEach(productDtoList::add);
             Map<String, Object> response = new HashMap<>();
-            response.put("products", products);
+            response.put("products", productDtoList);
             response.put("curPage", pageProducts.getNumber());
             response.put("totalPage", pageProducts.getTotalPages());
             response.put("totalElements", pageProducts.getTotalElements());
@@ -87,25 +113,53 @@ public class ProductService {
         }
     }
 
+    @Transactional
     public void deleteById(int id) {
         try {
-            productRepository.deleteById(id);
+            Optional<Product> productOpt = productRepository.findById(id);
+            if(!productOpt.isPresent()) {
+                throw new NoSuchElementException("Sản phẩm", "id", String.valueOf(id));
+            } else {
+                Product product = productOpt.get();
+                if(product.isDeleted()) {
+                    productRepository.deleteById(id);
+                } else {
+                    productRepository.updateDeletedById(id, true);
+                }
+            }
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Thực hiện không thành công. Vui lòng thử lại sau");
         }
     }
 
-    public Product updateProduct(int id, Product updatedProduct) {
+    @Transactional
+    public ProductDto updateProduct(int id, Product updatedProduct) {
         try {
             Optional<Product> optProduct = productRepository.findById(id);
             if(optProduct.isPresent()) {
                 Product product = optProduct.get();
                 updatedProduct.applyToProduct(product);
                 productRepository.save(product);
-                return product;
+                return product.toDto();
             } else {
                 throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Không tìm thấy danh mục có id " + id);
             }
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
+        }
+    }
+
+    @Transactional
+    public ProductDto updateSellStatus(int id, ProductDto productDto) {
+        try {
+            Optional<Product> product = productRepository.findById(id);
+            if(!product.isPresent()) {
+                throw new NoSuchElementException("Sản phẩm", "id", String.valueOf(id));
+            }
+            Product updatedProduct = product.get();
+            updatedProduct.setStopCell(productDto.isStopCell());
+            productRepository.save(updatedProduct);
+            return updatedProduct.toDto();
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
         }
